// types.ts
export enum CropType {
  WHEAT = 'Wheat',
  RICE = 'Rice',
  TOMATO = 'Tomato',
  POTATO = 'Potato',
  ONION = 'Onion',
  COTTON = 'Cotton',
  MAIZE = 'Maize'
}

export enum FarmerType {
  SMALL = 'Small/Marginal',
  MEDIUM = 'Medium',
  LARGE = 'Commercial/Large'
}

export type UserRole = 'farmer' | 'consumer' | null;

export interface UserInput {
  cropType: CropType;
  farmerType: FarmerType;
  harvestDate: string; 
  quantity: number; 
  quantityUnit: 'Kg' | 'Quintal' | 'Ton';
  location: string;
  storageAvailable: boolean;
  storageType?: 'Room' | 'Cold';
  storageDuration?: string;
  travelRange: number; // km
}